refactor: extract dashboard helpers
