+++ src/components/GraficaBarraSimple.tsx
@@ -1,0 +1,45 @@
+import React from "react";
+import { BarChart, Bar, PieChart, Pie, Cell, XAxis, YAxis, Tooltip, Legend, ResponsiveContainer } from "recharts";
+
+const colores = ["#538DD4", "#91CF50", "#FFFF00", "#FFA400", "#FF0000"];
+
+export default function GraficaBarraSimple({
+  resumen,
+  titulo,
+  chartType,
+}: {
+  resumen: any[];
+  titulo: string;
+  chartType: "bar" | "histogram" | "pie";
+}) {
+  return (
+    <div className="flex-1 min-h-[450px]">
+      <h4 className="font-bold mb-2 text-cogent-blue">{titulo}</h4>
+      <ResponsiveContainer width="100%" height={450}>
+        {chartType === "pie" ? (
+          <PieChart>
+            <Pie data={resumen} dataKey="cantidad" nameKey="nivel" label>
+              {resumen.map((_, i) => (
+                <Cell key={i} fill={colores[i % colores.length]} />
+              ))}
+            </Pie>
+            <Tooltip />
+            <Legend />
+          </PieChart>
+        ) : (
+          <BarChart data={resumen} barCategoryGap={chartType === "histogram" ? 0 : undefined}>
+            <XAxis dataKey="nivel" />
+            <YAxis allowDecimals={false} />
+            <Tooltip />
+            <Legend />
+            <Bar dataKey="cantidad" name="Cantidad">
+              {resumen.map((_, i) => (
+                <Cell key={i} fill={colores[i % colores.length]} />
+              ))}
+            </Bar>
+          </BarChart>
+        )}
+      </ResponsiveContainer>
+    </div>
+  );
+}
